data_ingestion/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile data_ingestion/requirements.in
#
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
cffi==1.14.2              # via google-crc32c
chardet==3.0.4            # via requests
google-api-core[grpc]==1.22.2  # via google-cloud-core, google-cloud-pubsub
google-auth==1.21.2       # via google-api-core, google-cloud-storage
google-cloud-core==1.4.1  # via -r data_ingestion/requirements.in, google-cloud-storage
google-cloud-pubsub==2.0.0  # via -r data_ingestion/requirements.in
google-cloud-storage==1.31.0  # via -r data_ingestion/requirements.in
google-crc32c==1.0.0      # via google-resumable-media
google-resumable-media==1.0.0  # via google-cloud-storage
googleapis-common-protos[grpc]==1.52.0  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.12.3  # via google-cloud-pubsub
grpcio==1.32.0            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
idna==2.10                # via requests
libcst==0.3.10            # via google-cloud-pubsub
mypy-extensions==0.4.3    # via typing-inspect
proto-plus==1.9.1         # via google-cloud-pubsub
protobuf==3.13.0          # via google-api-core, googleapis-common-protos, proto-plus
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pytz==2020.1              # via google-api-core
pyyaml==5.3.1             # via libcst
requests==2.24.0          # via -r data_ingestion/requirements.in, google-api-core
rsa==4.6                  # via google-auth
six==1.15.0               # via google-api-core, google-auth, google-resumable-media, grpcio, protobuf
typing-extensions==3.7.4.3  # via libcst, typing-inspect
typing-inspect==0.6.0     # via libcst
urllib3==1.25.10          # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
